Fixed & Enhanced!
Noticed there was an issue with some regex testing in the script (wasn't aware of POSIX standard of `\-` having to be at the start or end inside of square brackets).

Second issue I noticed was that depending on the Mediafire link format used, Mediafire will not use a normal button setup, so I needed to use a different means of searching by using regex again. (which eliminates having to use xml to find url).

Enhanced by simplifying the script into one file, and allowing for batch downloads.